+++ dsmr_frontend/static/dsmr_frontend/js/dsmr-reader/archive/electricity-returned.js
@@ -12,12 +12,13 @@ $(document).ready(function () {
 function render_electricity_returned_graph(xhr_data) {
     let echarts_options = {
         title: {
-            text: text_electricity_returned_header,
+            text: TEXT_ELECTRICITY_RETURNED_HEADER,
+            textStyle: TITLE_TEXTSTYLE,
             left: 'center'
         },
         color: [
-            electricity_returned_alternate_color,
-            electricity_returned_color
+            ELECTRICITY_RETURNED_ALTERNATE_COLOR,
+            ELECTRICITY_RETURNED_COLOR
         ],
         tooltip: {
             trigger: 'axis',
@@ -38,7 +39,7 @@ function render_electricity_returned_graph(xhr_data) {
         xAxis: [
             {
                 type: 'category',
-                boundaryGap: electricity_graph_style === 'bar',
+                boundaryGap: ELECTRICITY_GRAPH_STYLE === 'bar',
                 data: xhr_data.x
             }
         ],
@@ -53,17 +54,17 @@ function render_electricity_returned_graph(xhr_data) {
     if (xhr_data.electricity1_returned && xhr_data.electricity2_returned) {
         echarts_options.series = [
             {
-                name: text_electricity1_returned,
-                type: electricity_graph_style,
-                stack: stack_electricity_graphs,
+                name: TEXT_ELECTRICITY1_RETURNED,
+                type: ELECTRICITY_GRAPH_STYLE,
+                stack: STACK_ELECTRICITY_GRAPHS,
                 smooth: true,
                 areaStyle: {},
                 data: xhr_data.electricity1_returned
             },
             {
-                name: text_electricity2_returned,
-                type: electricity_graph_style,
-                stack: stack_electricity_graphs,
+                name: TEXT_ELECTRICITY2_RETURNED,
+                type: ELECTRICITY_GRAPH_STYLE,
+                stack: STACK_ELECTRICITY_GRAPHS,
                 smooth: true,
                 areaStyle: {},
                 data: xhr_data.electricity2_returned
@@ -72,8 +73,8 @@ function render_electricity_returned_graph(xhr_data) {
     } else if (xhr_data.electricity_returned_merged) {
         echarts_options.series = [
             {
-                name: text_electricity_merged_returned,
-                type: electricity_graph_style,
+                name: TEXT_ELECTRICITY_MERGED_RETURNED,
+                type: ELECTRICITY_GRAPH_STYLE,
                 smooth: true,
                 areaStyle: {},
                 data: xhr_data.electricity_returned_merged