+++ dsmr_frontend/static/dsmr_frontend/js/dsmr-reader/live/phases.js
@@ -1,15 +1,34 @@
 $(document).ready(function () {
 
     let echarts_phases_graph = echarts.init(document.getElementById('echarts-phases-graph'));
+    let x_axis = [
+        {
+            type: 'category',
+            boundaryGap: false,
+            data: null
+        },
+        {
+            type: 'category',
+            boundaryGap: false,
+            data: null,
+            inverse: true
+        }
+    ];
     let echarts_phases_initial_options = {
+        toolbox: TOOLBOX_OPTIONS,
         color: [
-            phase_delivered_l1_color,
-            phase_delivered_l2_color,
-            phase_delivered_l3_color,
-            phase_returned_l1_color,
-            phase_returned_l2_color,
-            phase_returned_l3_color
+            PHASE_DELIVERED_L1_COLOR,
+            PHASE_DELIVERED_L2_COLOR,
+            PHASE_DELIVERED_L3_COLOR,
+            PHASE_RETURNED_L1_COLOR,
+            PHASE_RETURNED_L2_COLOR,
+            PHASE_RETURNED_L3_COLOR
         ],
+        title: {
+            text: TEXT_PHASES_HEADER,
+            textStyle: TITLE_TEXTSTYLE,
+            left: 'center',
+        },
         tooltip: {
             trigger: 'axis',
             axisPointer: {
@@ -26,22 +45,20 @@ $(document).ready(function () {
             right: '2%',
             containLabel: true
         },
-        xAxis: [
-            {
-                type: 'category',
-                boundaryGap: false,
-                data: null
-            }
-        ],
+        xAxis: x_axis,
         yAxis: [
             {
                 type: 'value'
+            },
+            {
+                type: 'value',
+                inverse: true
             }
         ],
         dataZoom: [
             {
                 show: true,
-                start: live_graphs_initial_zoom,
+                start: LIVE_GRAPHS_INITIAL_ZOOM,
                 end: 100
             },
             {
@@ -54,16 +71,11 @@ $(document).ready(function () {
 
     /* These settings should not affect the updates and reset the zoom on each update. */
     let echarts_phases_update_options = {
-        xAxis: [
-            {
-                type: 'category',
-                boundaryGap: false,
-                data: null
-            }
-        ],
+        xAxis: x_axis,
         series: [
             {
-                name: 'Watt (L1)',
+                yAxisIndex: 0,
+                name: 'L1+',
                 type: 'line',
                 smooth: true,
                 areaStyle: {},
@@ -71,7 +83,8 @@ $(document).ready(function () {
                 data: null
             },
             {
-                name: 'Watt (L2)',
+                yAxisIndex: 0,
+                name: 'L2+',
                 type: 'line',
                 smooth: true,
                 areaStyle: {},
@@ -79,7 +92,8 @@ $(document).ready(function () {
                 data: null
             },
             {
-                name: 'Watt (L3)',
+                yAxisIndex: 0,
+                name: 'L3+',
                 type: 'line',
                 smooth: true,
                 areaStyle: {},
@@ -87,7 +101,8 @@ $(document).ready(function () {
                 data: null
             },
             {
-                name: 'Watt (L1)',
+                yAxisIndex: 1,
+                name: 'L1-',
                 type: 'line',
                 smooth: true,
                 areaStyle: {},
@@ -95,7 +110,8 @@ $(document).ready(function () {
                 data: null
             },
             {
-                name: 'Watt (L2)',
+                yAxisIndex: 1,
+                name: 'L2-',
                 type: 'line',
                 smooth: true,
                 areaStyle: {},
@@ -103,7 +119,8 @@ $(document).ready(function () {
                 data: null
             },
             {
-                name: 'Watt (L3)',
+                yAxisIndex: 1,
+                name: 'L3-',
                 type: 'line',
                 smooth: true,
                 areaStyle: {},
@@ -113,15 +130,16 @@ $(document).ready(function () {
         ]
     };
 
-    echarts_phases_graph.showLoading('default', echarts_loading_options);
+    echarts_phases_graph.showLoading('default', ECHARTS_LOADING_OPTIONS);
 
     /* Init graph. */
-    $.get(echarts_phases_graph_url, function (xhr_data) {
+    $.get(ECHARTS_PHASES_GRAPH_URL, function (xhr_data) {
         echarts_phases_graph.hideLoading();
         echarts_phases_graph.setOption(echarts_phases_initial_options);
 
         /* Different set of options, to prevent the dataZoom being reset on each update. */
         echarts_phases_update_options.xAxis[0].data = xhr_data.read_at;
+        echarts_phases_update_options.xAxis[1].data = xhr_data.read_at;
         echarts_phases_update_options.series[0].data = xhr_data.phases_delivered.l1;
         echarts_phases_update_options.series[1].data = xhr_data.phases_delivered.l2;
         echarts_phases_update_options.series[2].data = xhr_data.phases_delivered.l3;
@@ -143,7 +161,7 @@ $(document).ready(function () {
 
             pending_xhr_request = $.ajax({
                 dataType: "json",
-                url: echarts_phases_graph_url + "&latest_delta_id=" + latest_delta_id,
+                url: ECHARTS_PHASES_GRAPH_URL + "&latest_delta_id=" + latest_delta_id,
             }).done(function(xhr_data) {
                 /* Ignore empty sets. */
                 if (xhr_data.read_at.length === 0) {
@@ -153,6 +171,7 @@ $(document).ready(function () {
                 /* Delta update. */
                 for (let i = 0; i < xhr_data.read_at.length; i++) {
                     echarts_phases_update_options.xAxis[0].data.push(xhr_data.read_at[i]);
+                    echarts_phases_update_options.xAxis[1].data.push(xhr_data.read_at[i]);
                     echarts_phases_update_options.series[0].data.push(xhr_data.phases_delivered.l1[i]);
                     echarts_phases_update_options.series[1].data.push(xhr_data.phases_delivered.l2[i]);
                     echarts_phases_update_options.series[2].data.push(xhr_data.phases_delivered.l3[i]);
@@ -167,7 +186,7 @@ $(document).ready(function () {
                 // Allow new updates
                 pending_xhr_request = null;
             });
-        }, echarts_phases_graph_interval * 1000);
+        }, ECHARTS_PHASES_GRAPH_INTERVAL * 1000);
     });
 
     /* Responsiveness. */