+++ dsmr_frontend/static/dsmr_frontend/js/dsmr-reader/live/power-current.js
@@ -1,11 +1,17 @@
 $(document).ready(function () {
     let echarts_power_current_graph = echarts.init(document.getElementById('echarts-power-current-graph'));
     let echarts_power_current_initial_options = {
+        toolbox: TOOLBOX_OPTIONS,
         color: [
-            power_current_l1_color,
-            power_current_l2_color,
-            power_current_l3_color
+            POWER_CURRENT_L1_COLOR,
+            POWER_CURRENT_L2_COLOR,
+            POWER_CURRENT_L3_COLOR
         ],
+        title: {
+            text: TEXT_POWER_CURRENT_HEADER,
+            textStyle: TITLE_TEXTSTYLE,
+            left: 'center',
+        },
         tooltip: {
             trigger: 'axis',
             axisPointer: {
@@ -39,7 +45,7 @@ $(document).ready(function () {
         dataZoom: [
             {
                 show: true,
-                start: live_graphs_initial_zoom,
+                start: LIVE_GRAPHS_INITIAL_ZOOM,
                 end: 100
             },
             {
@@ -62,29 +68,29 @@ $(document).ready(function () {
         series: null
     };
 
-    echarts_power_current_graph.showLoading('default', echarts_loading_options);
+    echarts_power_current_graph.showLoading('default', ECHARTS_LOADING_OPTIONS);
 
     /* Init graph. */
-    $.get(echarts_power_current_graph_url, function (xhr_data) {
+    $.get(ECHARTS_POWER_CURRENT_GRAPH_URL, function (xhr_data) {
         echarts_power_current_graph.hideLoading();
 
         /* Dynamic phases. */
-        if (is_multi_phase) {
+        if (IS_MULTI_PHASE) {
             echarts_power_current_update_options.series = [
                 {
-                    name: 'Ampere (L1)',
+                    name: 'L1',
                     type: 'bar',
                     stack: true,
                     data: null
                 },
                 {
-                    name: 'Ampere (L2)',
+                    name: 'L2',
                     type: 'bar',
                     stack: true,
                     data: null
                 },
                 {
-                    name: 'Ampere (L3)',
+                    name: 'L3',
                     type: 'bar',
                     stack: true,
                     data: null
@@ -106,7 +112,7 @@ $(document).ready(function () {
         echarts_power_current_update_options.xAxis[0].data = xhr_data.read_at;
         echarts_power_current_update_options.series[0].data = xhr_data.phase_power_current.l1;
 
-        if (is_multi_phase) {
+        if (IS_MULTI_PHASE) {
             echarts_power_current_update_options.series[1].data = xhr_data.phase_power_current.l2;
             echarts_power_current_update_options.series[2].data = xhr_data.phase_power_current.l3;
         }
@@ -126,7 +132,7 @@ $(document).ready(function () {
 
             pending_xhr_request = $.ajax({
                 dataType: "json",
-                url: echarts_power_current_graph_url + "&latest_delta_id=" + latest_delta_id,
+                url: ECHARTS_POWER_CURRENT_GRAPH_URL + "&latest_delta_id=" + latest_delta_id,
             }).done(function(xhr_data) {
                 /* Ignore empty sets. */
                 if (xhr_data.read_at.length === 0) {
@@ -138,7 +144,7 @@ $(document).ready(function () {
                     echarts_power_current_update_options.xAxis[0].data.push(xhr_data.read_at[i]);
                     echarts_power_current_update_options.series[0].data.push(xhr_data.phase_power_current.l1[i]);
 
-                    if (is_multi_phase) {
+                    if (IS_MULTI_PHASE) {
                         echarts_power_current_update_options.series[1].data.push(xhr_data.phase_power_current.l2[i]);
                         echarts_power_current_update_options.series[2].data.push(xhr_data.phase_power_current.l3[i]);
                     }
@@ -150,7 +156,7 @@ $(document).ready(function () {
                 // Allow new updates
                 pending_xhr_request = null;
             });
-        }, echarts_power_current_graph_interval * 1000);
+        }, ECHARTS_POWER_CURRENT_GRAPH_INTERVAL * 1000);
     });
 
     /* Responsiveness. */